Refactor code and remove unnecessary packages
This commit replaces the direct assignments of StressRecord class with constructor initialization. It also removes an unused 'freezed_annotation' package from the nitrite_demo's pubspec.yaml file and the corresponding import statements. The unused lines of code commented in entity_parser.dart file are also removed. The nitrite_entity_generator path is replaced with the version in the pubspec.yaml files. Refactored the creation of the 'Todo' instances and moved the Freezed package from 'dev_dependencies' to 'dependencies' in nitrite_hive_adapter.

This refactoring enhances the consistency and readability of the code and also optimizes the packages used. It ensures the nitrite entity generator is referenced correctly in the project.